refactor(back): Move function to a sound place.
